pyproject.toml
[tool.poetry]
name = "discord-key-bot"
version = "0.1.0"
description = ""
authors = ["Ryan Barnes <ryan.bayangan@gmail.com>"]
readme = "README.md"
packages = [{include = "discord_key_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
typer = "*"
sqlmodel = "*"
discord-py = "^2.3.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.272"
mypy = "^1.3.0"

[tool.ruff]
select = [
  "F", # Pyflakes
  "E", # Pycodestyle
  "W", # Pycodestyle

  #"C90", # Mccabe complexity
  #"C",   # flake8-comprehensions
  #"B",   # flake8-bugbear
  "UP", # pyupgrade
  "S",  # bandit (security)
  #"DTZ", # Enforce good datetime
  "I", # isort
]


ignore = [
  "E501", # Line length is done by black
  # "D202", # NoBlankLineAfterFunction
  "S101", # Use of `assert` detected
  "S104", # Ignore binding to all interfaces.
  "S110", # try - except - pass. TODO: Remove this ignore
  "S113", # requests call without timeout. TODO: Remove this ignore
  "S602", # remove shell=True... not sure how we will do this
  # TODO: WE must fix these xss vulnerabilities.
  "S308", # mark_safe. We have lots of XSS vulns here but too many to fix in this pr
]

# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py310"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
